+++ src/app/logs/page.tsx
@@ -6,10 +6,10 @@ import Link from 'next/link';
 import { useAtomValue } from 'jotai';
 import { atom_status } from '@/atoms';
 
-import UserTabs from '@/components/sections/UserTabs';
-import ChatLogs from '@/components/sections/ChatLogs';
+import Warn from '@/components/shared/Warn';
 
-import Warn from '@/components/parts/Warn';
+import UserTabs from './UserTabs';
+import ChatLogs from './ChatLogs';
 
 export default function Page() {
   const status = useAtomValue(atom_status);
