+++ src/app/not-found.tsx
@@ -1,4 +1,4 @@
-import Warn from '@/components/parts/Warn';
+import Warn from '@/components/shared/Warn';
 
 export default function NotFound() {
   return <Warn>404 | This page could not be found.</Warn>;
