+++ src/app/spy/Command.tsx
@@ -2,18 +2,19 @@
 
 import { useState } from 'react';
 
-import Area from '@/components/wrappers/Area';
-import Label from '@/components/parts/Label';
-import Input from '@/components/parts/Input';
-import InputError from '@/components/parts/InputError';
-import ControlBtn from '@/components/parts/ControlBtn';
-
 import { useAtom, useAtomValue } from 'jotai';
 import { atom_users, atom_channels, atom_status, atom_isLoading, atom_idbConn, atom_tmiConn } from '@/atoms';
 
 import tmi from 'tmi.js';
 import { deleteDB, openDB } from 'idb';
 import Chat2Db from '@/lib/Chat2Db';
+
+import SectionArea from '@/components/layout/SectionArea';
+
+import Label from '@/components/Label';
+import Input from '@/components/Input';
+import InputError from '@/components/InputError';
+import ControlBtn from '@/components/ControlBtn';
 
 export default function Command() {
   const [isError, setIsError] = useState({
@@ -131,7 +132,7 @@ export default function Command() {
   // }
 
   return (
-    <Area.Section title="Command">
+    <SectionArea title="Command">
       <Label title="Users:" htmlFor="users" desc="the users you are going to track">
         <Input id="users" placeholder="Enter usernames separated with spaces." itemsAtom={atom_users} />
         {isError.users && <InputError message="Please enter at least one user." />}
@@ -147,6 +148,6 @@ export default function Command() {
           <ControlBtn variant="start" isDisabled={isLoading} onClick={handleStart} />
         )}
       </div>
-    </Area.Section>
+    </SectionArea>
   );
 }
