+++ src/components/LogsInfo.tsx
