--- src/components/parts/SectionTitle.tsx
@@ -1,3 +1,0 @@
-export default function SectionTitle({ children }: { children: React.ReactNode }) {
-  return <h2 className="border-b border-c_border2 pb-2 text-sm font-medium text-gray-300 lg:text-base">{children}</h2>;
-}
