+++ src/components/shared/icons/IconChevron.tsx
@@ -1,9 +1,7 @@
-function IconChevron({ className }: { className?: string }) {
+export default function IconChevron({ className }: { className: string }) {
   return (
     <svg viewBox="0 0 14 24" fill="currentColor" height="1em" width="1em" className={className}>
       <path d="M8.59 16.58L13.17 12 8.59 7.41 10 6l6 6-6 6-1.41-1.42z" />
     </svg>
   );
 }
-
-export default IconChevron;
