+++ src/components/shared/icons/IconHome.tsx
@@ -1,4 +1,4 @@
-function IconStop({ className }: { className?: string }) {
+export default function IconHome({ className }: { className: string }) {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
@@ -7,14 +7,13 @@ function IconStop({ className }: { className?: string }) {
       viewBox="0 0 24 24"
       fill="none"
       stroke="currentColor"
-      strokeWidth="4"
+      strokeWidth="2"
       strokeLinecap="round"
       strokeLinejoin="round"
       className={className}
     >
-      <rect width="18" height="18" x="3" y="3" rx="2" />
+      <path d="m3 9 9-7 9 7v11a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2z" />
+      <polyline points="9 22 9 12 15 12 15 22" />
     </svg>
   );
 }
-
-export default IconStop;
