+++ src/components/shared/icons/IconStart.tsx
@@ -1,4 +1,4 @@
-function IconSpy({ className }: { className?: string }) {
+export default function IconStart({ className }: { className: string }) {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
@@ -7,17 +7,12 @@ function IconSpy({ className }: { className?: string }) {
       viewBox="0 0 24 24"
       fill="none"
       stroke="currentColor"
-      strokeWidth="2"
+      strokeWidth="4"
       strokeLinecap="round"
       strokeLinejoin="round"
       className={className}
     >
-      <path d="M4 10a7.31 7.31 0 0 0 10 10Z" />
-      <path d="m9 15 3-3" />
-      <path d="M17 13a6 6 0 0 0-6-6" />
-      <path d="M21 13A10 10 0 0 0 11 3" />
+      <polygon points="5 3 19 12 5 21 5 3" />
     </svg>
   );
 }
-
-export default IconSpy;
