+++ src/components/shared/icons/IconStop.tsx
@@ -1,4 +1,4 @@
-function IconHome({ className }: { className?: string }) {
+export default function IconStop({ className }: { className: string }) {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
@@ -7,15 +7,12 @@ function IconHome({ className }: { className?: string }) {
       viewBox="0 0 24 24"
       fill="none"
       stroke="currentColor"
-      strokeWidth="2"
+      strokeWidth="4"
       strokeLinecap="round"
       strokeLinejoin="round"
       className={className}
     >
-      <path d="m3 9 9-7 9 7v11a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2z" />
-      <polyline points="9 22 9 12 15 12 15 22" />
+      <rect width="18" height="18" x="3" y="3" rx="2" />
     </svg>
   );
 }
-
-export default IconHome;
