+++ src/components/zone/Header.tsx
@@ -1,15 +1,14 @@
-import Container from '@/components/wrappers/Container';
+import Container from '@/components/layout/Container';
 
-import Logo from '@/components/parts/Logo';
-import GithubBtn from '@/components/parts/GithubBtn';
+import HeaderBtn from '@/components/HeaderBtn';
 
-export function Header() {
+export default function Header() {
   return (
     <header className="z-20 border-b border-c_border1 bg-c_header sm:sticky sm:top-0">
       <Container>
         <div className="flex items-center justify-between">
-          <Logo />
-          <GithubBtn />
+          <HeaderBtn.Logo />
+          <HeaderBtn.GitHub />
         </div>
       </Container>
     </header>
