+++ src/components/zone/Statusbar.tsx
@@ -7,16 +7,7 @@ import { atom_status } from '@/atoms';
 
 import { formatMilliseconds } from '@/lib/utils';
 
-function Info({ value, label }: { value: string; label: string }) {
-  return (
-    <div className="flex min-w-max flex-1 flex-row-reverse items-center justify-end gap-1 overflow-hidden px-2 pb-2 pt-2 sm:flex-col sm:items-stretch sm:justify-stretch sm:gap-0 sm:px-4 sm:pt-0">
-      <span className="w-0 text-sm font-medium text-gray-300 xl:text-base" suppressHydrationWarning={true}>
-        {value}
-      </span>
-      <span className="text-xs text-gray-400 xl:text-sm">{label}</span>
-    </div>
-  );
-}
+import StatusInfo from '@/components/StatusInfo';
 
 export default function Statusbar() {
   const status = useAtomValue(atom_status);
@@ -45,9 +36,9 @@ export default function Statusbar() {
     <div className="overflow-hidden rounded-xl border border-c_border1 bg-c_main">
       <div className="border-c_border2 sm:pt-3 lg:border-b lg:bg-c_body xl:pt-5">
         <div className="flex divide-x divide-c_border1 lg:divide-c_border2">
-          <Info value={status.uCount.toString()} label="Users" />
-          <Info value={status.cCount.toString()} label="Channels" />
-          <Info value={uptime} label="Up Time" />
+          <StatusInfo value={status.uCount.toString()} label="Users" />
+          <StatusInfo value={status.cCount.toString()} label="Channels" />
+          <StatusInfo value={uptime} label="Up Time" />
         </div>
       </div>
       <div className="hidden py-4 text-center text-sm text-gray-300 lg:block xl:py-8 xl:text-base">
